Refresh schematic render when F3+A is pressed
This debug shortcut (which is not rebindable in vanilla, and thus is hardcoded here) is used to reload/rerender all chunks.  It's useful to also be able to reload the content in schematics.  While for the average user this shouldn't be necessary, while debuging it can be nice (and sometimes a situation pops up when you want to reload it).

I surrounded the code in question with an instanceof check because otherwise it would access the last keyboard event during mouse events.  That would be bad, but there's also MinecraftForge/MinecraftForge#4270 which causes the mouse event to fire each tick instead of when the mouse actually moves - basically, without it there would be a second, unwanted rerender.